Initial refactor into task modules, to remove dependency on importing dags/ modules in integration tests
