+++ makerules/python.mk
@@ -13,16 +13,13 @@ black:
 flake8:
 	flake8 .
 
-test:: test-unit test-integration test-e2e
+test:: test-unit test-integration
 
 test-unit:
-	[ -d digital_land_airflow/tests/unit ] && python -m pytest digital_land_airflow/tests/unit --junitxml=.junitxml/unit.xml -c setup.cfg || true
+	[ -d digital_land_airflow/tests/unit ] && python -m pytest digital_land_airflow/tests/unit --junitxml=.junitxml/unit.xml -c setup.cfg $(PYTEST_EXTRA_ARGS)
 
 test-integration:
-	[ -d digital_land_airflow/tests/integration ] && python -m pytest digital_land_airflow/tests/integration --junitxml=.junitxml/integration.xml -c setup.cfg
-
-test-e2e:
-	[ -d digital_land_airflow/tests/e2e ] && python -m pytest digital_land_airflow/tests/e2e --junitxml=.junitxml/e2e.xml -c setup.cfg || true
+	[ -d digital_land_airflow/tests/integration ] && python -m pytest digital_land_airflow/tests/integration --junitxml=.junitxml/integration.xml -c setup.cfg $(PYTEST_EXTRA_ARGS)
 
 coverage:
 	coverage run --source $(PACKAGE) -m py.test && coverage report
